Add note to myself.
I think this is good enough for now.  It will catch obvious mistakes in
parameter ordering.  I'm not sure if it will work with all situations,
like doing an orth_proj on D, B, C, S, A, S  to S, A, S.  But these
situations aren't really relevant to this application at the moment.

Closes #13